// src/types/recruitingEvaluation.ts
export interface RecruitingClassScoreDistribution {
  teams: number;
  minimum: number;
  p10: number;
  p25: number;
  median: number;
  p75: number;
  p90: number;
  maximum: number;
  mean: number;
  standardDeviation: number;
  exactDistinctScores: number;
  displayedDistinctScores: number;
  exactTieRate: number;
  displayedTieRate: number;
}

export interface RecruitingCountDistribution {
  count: number;
  minimum: number;
  p10: number;
  p25: number;
  median: number;
  p75: number;
  p90: number;
  maximum: number;
  mean: number;
}

export interface RecruitingSupplySummary {
  available: number;
  signed: number;
  unsigned: number;
  signingRate: number;
}

export interface RecruitingTop25PrestigeEntry {
  eligibleTeamSeasons: number;
  appearances: number;
  compositionShare: number;
  appearanceRate: number;
}

export type RecruitingTop25Composition = Record<
  number,
  RecruitingTop25PrestigeEntry
>;
// ...
export interface RecruitingEvaluationAggregate {
  signingDayShare: number;
  baseCapacityCompletion: number;
  teamsCompletingBaseCapacity: number;
  teamBaseCapacityCompletionRate: number;
  oversigningsPerTeamSeason: number;
  teamsUsingAllFourOversigns: number;
  teamsUsingAllFourOversignsRate: number;
  walkOnsPerTeamSeason: number;
  teamSeasonsUsingWalkOns: number;
  teamSeasonsUsingWalkOnsRate: number;
  meaningfulPursuits: number;
  meaningfullyPursuedProspects: number;
  contestedMeaningfulProspects: number;
  meaningfulCompetitionRate: number;
  pursuitsAdmitted: number;
  fundableOpeningsUnfilled: number;
  lowPrestigeEliteWins: number;
  lowPrestigeEliteShare: number;
  prestigeClassScoreCorrelation: number;
  classScoreDistribution: RecruitingClassScoreDistribution;
  classSizeDistribution: RecruitingCountDistribution;
  supplyByStar: Record<number, RecruitingSupplySummary>;
  supplyByPosition: Record<string, RecruitingSupplySummary>;
  top25ClassComposition: RecruitingTop25Composition;
  ratingSpreadChange: number;
  steadyStateRatingSpreadChange: number;
  prestigeMobility: number;
  prestigeMobilityRate: number;
}